# packages/assistant-api-contract/openapi.json
{
  "components": {
    "schemas": {
      "ActionApprovalRequest": {
        "properties": {
          "actor": {
            "default": "user_demo",
            "title": "Actor",
            "type": "string"
          },
          "channel": {
            "$ref": "#/components/schemas/ApprovalChannel",
            "default": "web"
          },
          "fresh_auth": {
            "default": false,
            "title": "Fresh Auth",
            "type": "boolean"
          }
        },
        "title": "ActionApprovalRequest",
        "type": "object"
      },
      "ActionCreateRequest": {
        "properties": {
          "action_type": {
            "default": "create_email_draft",
            "title": "Action Type",
            "type": "string"
          },
          "changed_fields": {
            "items": {
              "type": "string"
            },
            "title": "Changed Fields",
            "type": "array"
          },
          "external_side_effect": {
            "default": true,
            "title": "External Side Effect",
            "type": "boolean"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "recipient_refs": {
            "items": {
              "type": "string"
            },
            "title": "Recipient Refs",
            "type": "array"
          },
          "reversible": {
            "default": true,
            "title": "Reversible",
            "type": "boolean"
          },
          "risk_tier": {
            "$ref": "#/components/schemas/RiskTier",
            "default": "medium"
          },
          "scope": {
            "$ref": "#/components/schemas/ScopedIdentity"
          },
          "sending_account_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "onebrain://connected-account/demo-mail",
            "title": "Sending Account Ref"
          },
          "sensitive_flags": {
            "items": {
              "type": "string"
            },
            "title": "Sensitive Flags",
            "type": "array"
          },
          "source_refs": {
            "items": {
              "type": "string"
            },
            "title": "Source Refs",
            "type": "array"
          },
          "summary": {
            "default": "Draft a reply for review.",
            "title": "Summary",
            "type": "string"
          }
        },
        "title": "ActionCreateRequest",
        "type": "object"
      },
      "ActionRecord": {
        "properties": {
          "action_id": {
            "format": "uuid",
            "title": "Action Id",
            "type": "string"
          },
          "action_type": {
            "title": "Action Type",
            "type": "string"
          },
          "approval_reason": {
            "default": "User approval required by action policy.",
            "title": "Approval Reason",
            "type": "string"
          },
          "audit_correlation_id": {
            "title": "Audit Correlation Id",
            "type": "string"
          },
          "changed_fields": {
            "items": {
              "type": "string"
            },
            "title": "Changed Fields",
            "type": "array"
          },
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "external_side_effect": {
            "default": true,
            "title": "External Side Effect",
            "type": "boolean"
          },
          "idempotency_key": {
            "title": "Idempotency Key",
            "type": "string"
          },
          "recipient_refs": {
            "items": {
              "type": "string"
            },
            "title": "Recipient Refs",
            "type": "array"
          },
          "reversible": {
            "default": false,
            "title": "Reversible",
            "type": "boolean"
          },
          "risk_tier": {
            "$ref": "#/components/schemas/RiskTier"
          },
          "scope": {
            "$ref": "#/components/schemas/ScopedIdentity"
          },
          "sending_account_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sending Account Ref"
          },
          "sensitive_flags": {
            "items": {
              "type": "string"
            },
            "title": "Sensitive Flags",
            "type": "array"
          },
          "source_refs": {
            "items": {
              "type": "string"
            },
            "title": "Source Refs",
            "type": "array"
          },
          "state": {
            "$ref": "#/components/schemas/ActionState",
            "default": "proposed"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "transitions": {
            "items": {
              "$ref": "#/components/schemas/TransitionRecord"
            },
            "title": "Transitions",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "scope",
          "action_type",
          "risk_tier",
          "summary",
          "idempotency_key"
        ],
        "title": "ActionRecord",
        "type": "object"
      },
      "ActionState": {
        "enum": [
          "proposed",
          "needs_review",
          "approved",
          "executing",
          "executed",
          "failed",
          "cancelled"
        ],
        "title": "ActionState",
        "type": "string"
      },
...
      "ApprovalChannel": {
        "enum": [
          "web",
          "telegram",
          "voice",
          "worker",
          "fresh_auth"
        ],
        "title": "ApprovalChannel",
        "type": "string"
      },
...
      "FirewallDecision": {
        "properties": {
          "blocked_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Blocked Reasons",
            "type": "array"
          },
          "can_create_action_from_raw_content": {
            "default": false,
            "title": "Can Create Action From Raw Content",
            "type": "boolean"
          },
          "extracted_intent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extracted Intent"
          },
          "proposed_recipient_refs": {
            "items": {
              "type": "string"
            },
            "title": "Proposed Recipient Refs",
            "type": "array"
          },
          "risk_flags": {
            "items": {
              "type": "string"
            },
            "title": "Risk Flags",
            "type": "array"
          },
          "safe_for_model_context": {
            "title": "Safe For Model Context",
            "type": "boolean"
          }
        },
        "required": [
          "safe_for_model_context"
        ],
        "title": "FirewallDecision",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RiskTier": {
        "enum": [
          "low",
          "medium",
          "high"
        ],
        "title": "RiskTier",
        "type": "string"
      },
      "SanitizedContent": {
        "properties": {
          "removed_markers": {
            "items": {
              "type": "string"
            },
            "title": "Removed Markers",
            "type": "array"
          },
          "risk_flags": {
            "items": {
              "type": "string"
            },
            "title": "Risk Flags",
            "type": "array"
          },
          "safe_text": {
            "title": "Safe Text",
            "type": "string"
          }
        },
        "required": [
          "safe_text"
        ],
        "title": "SanitizedContent",
        "type": "object"
      },
      "ScopedIdentity": {
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "purpose": {
            "default": "assistant_operations",
            "title": "Purpose",
            "type": "string"
          },
          "space_id": {
            "title": "Space Id",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "account_id",
          "user_id",
          "space_id"
        ],
        "title": "ScopedIdentity",
        "type": "object"
      },
      "SecurityInspectionRequest": {
        "properties": {
          "html": {
            "title": "Html",
            "type": "string"
          },
          "source_ref": {
            "default": "onebrain://fixture/untrusted",
            "title": "Source Ref",
            "type": "string"
          }
        },
        "required": [
          "html"
        ],
        "title": "SecurityInspectionRequest",
        "type": "object"
      },
      "SecurityInspectionResponse": {
        "properties": {
          "firewall": {
            "$ref": "#/components/schemas/FirewallDecision"
          },
          "sanitized": {
            "$ref": "#/components/schemas/SanitizedContent"
          }
        },
        "required": [
          "sanitized",
          "firewall"
        ],
        "title": "SecurityInspectionResponse",
        "type": "object"
      },
...
      "TransitionRecord": {
        "properties": {
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "channel": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApprovalChannel"
              },
              {
                "type": "null"
              }
            ]
          },
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "from_state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActionState"
              },
              {
                "type": "null"
              }
            ]
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "to_state": {
            "$ref": "#/components/schemas/ActionState"
          }
        },
        "required": [
          "from_state",
          "to_state",
          "actor",
          "reason",
          "correlation_id"
        ],
        "title": "TransitionRecord",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Thin orchestration API for the OneBrain assistant.",
    "title": "OneBrain Assistant API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/actions": {
      "post": {
        "operationId": "create_action_v1_actions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Action"
      }
    },
    "/v1/actions/{action_id}/approve": {
      "post": {
        "operationId": "approve_action_v1_actions__action_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Action Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Action"
      }
    },
    "/v1/security/inspect": {
      "post": {
        "operationId": "inspect_untrusted_content_v1_security_inspect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SecurityInspectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecurityInspectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Inspect Untrusted Content"
      }
    },
    "/v1/telegram/webhook": {
      "post": {
        "operationId": "telegram_webhook_v1_telegram_webhook_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Telegram Webhook V1 Telegram Webhook Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Telegram Webhook"
      }
    },
...
  }
}